feat: default to Precision.fp32 and pin flutter_litert ^3.8.0
Adopts flutter_litert 3.8.0 and follows its precision default change.

fp16 was the default at 4 call sites. Across 29 published detection models
measured on five GPUs, fp16 matched a plain-CPU reference for only about a
fifth of them while fp32 matched every model that compiled. Detection outputs
are pixel-space box coordinates, and fp16 carries roughly three decimal digits
of mantissa, so the error lands directly on box geometry.

The cost is real rather than free: fp32 is a median 29.9% slower on GPU across
those five GPUs, with Apple M4 the lone exception at 6.5% faster. Every changed
site was a default parameter value, never an explicit call-site choice, so the
parameter is retained and callers can pass Precision.fp16 to restore the old
behaviour.

Folded into the pending 0.4.0, which is already a minor bump over the published
0.3.0 and has not shipped.

+++ lib/src/isolate/object_detector_core.dart
@@ -64,7 +64,7 @@ class _ObjectDetectorCore {
     PerformanceConfig performanceConfig = const PerformanceConfig(),
     bool useCompiledModel = false,
     Set<Accelerator> accelerators = const {Accelerator.gpu, Accelerator.cpu},
-    Precision precision = Precision.fp16,
+    Precision precision = Precision.fp32,
   }) async {
     try {
       _model = useCompiledModel
